Versão Final (entregue)
Versão que foi entregue

+++ JS/scriptCad.js
@@ -1,10 +1,12 @@
 $(document).ready(function () {
+    // Máscaras de entrada
     $("#cpf").mask("000.000.000-00");
     $("#rg").mask("00.000.000-0");
     $("#telefone").mask("(00) 00000-0000");
     $('#cartaoCredito').mask('0000 0000 0000 0000');
     $('#cartaoDebito').mask('0000 0000 0000 0000');
 
+    // Função de cadastro
     $("#cadastroForm").on("submit", function (event) {
         event.preventDefault();
 
@@ -32,24 +34,38 @@ $(document).ready(function () {
             chavePix: chavePix
         };
 
-        localStorage.setItem("cliente", JSON.stringify(cliente));
-        alert("Cliente cadastrado com sucesso!");
-        window.location.href = "login.html";
+        // Verificação de cliente existente
+        var clientes = JSON.parse(localStorage.getItem('clientes')) || [];
+        var clienteExistente = clientes.find(c => c.cpf === cliente.cpf || c.email === cliente.email);
+
+        
+        if (clienteExistente) {
+            alert('Já existe um cliente cadastrado com esse CPF ou E-mail.');
+        } else {
+            clientes.push(cliente);
+            localStorage.setItem('clientes', JSON.stringify(clientes));
+            localStorage.setItem("cliente", JSON.stringify(cliente));
+            alert("Cliente cadastrado com sucesso!");
+            window.location.href = "login.html";
+        }
     });
 
+    // Função de logout
     $("#btLogout").click(function () {
-        localStorage.removeItem("loggedIn");
-        updateNavbar();
-        alert('Perfil Deslogado com sucesso!');
+        localStorage.removeItem("logado");
+        attMenu();
+        alert('Perfil deslogado com sucesso!');
         window.location.href = "login.html";
     });
 
-    updateNavbar();
+    // Atualização do menu
+    attMenu();
 });
 
-function updateNavbar() {
-    var isLoggedIn = localStorage.getItem("loggedIn");
-    if (isLoggedIn) {
+// Função para atualizar o menu de navegação
+function attMenu() {
+    var estaLogado = localStorage.getItem("logado");
+    if (estaLogado) {
         $("#perfilLink").show();
         $("#loginId").hide();
         $("#btLogout").show();
